backend/requirements.txt
fastapi
uvicorn

#for database
sqlalchemy
psycopg2
pymysql

# for env var
python-dotenv